Updated diary, changes csv, updated test1 and created test2 for model var approach
I updated my diary, I removed 2 days of stocks from the csv to make it so there are 500 days total, an easier number to find a percentile for, I updated test.py so that it calculates percentage change between each day, and finds a percentile of that. Finally, I created test2 that connects to yahoo via an api and it uses variance covariance method to generate a var, using the same data as test.