Creating Interim Report with full contents page, and updated Diary...
For the foreseeable future, with each commit, it can be generally assumed that the diary has been updated alongside the commit to some capacity, so I will stop mentioning it in the title. I copied the previous project plan, adjusted it to be my interim report and created all the necessary headings for the contents page.