Implemented Back-Testing
After reading up on back testing, I decided to attempt to implement it for historical simulation (test1), it performs a rolling window of 100 days and checks the VaR at the end of each 100 days, if it exceeds the VaR loss then its added to a count, then the count is divided into a percentage based on the amount of windows used, seeing how accurate the VaR seems to be. At the 5% level, it appears my VaR is. I also added to Diary.