Adjusted Current Stock Background and Added a working Current Ino
I made it so that the current stock background looked better then just the pure black it was, and gave it a border and made it the same size as the search to its left. I also Added working current info, that whenever info was entered and you click the enter key (which wasn't working as I had not set multiline to false), it will delete everything and redraw it with updated details, to make sure that the variables are saving correctly and also so the user knows what they have entered.