Created working radio buttons for VaR method, removed right scroll view
I used ToggleButtons so that I can have radio buttons working for what method I want to use for VaR. It does this in reference to a saved variable, that will be updated depending on which button is pressed, so I will be able to refer to this in the future. I also decided to make the right hand section non-scrollable since every bit of information was able to fit there without any problems. Next I will change its background and try starting Unit Testing soon