Added a background and fixed a bug
Added a gray background to the inputs to help them be differentiated, its all quite poor aesthetically but I will work on that in the future. I also fixed a big where the radio buttons would reset to the wrong one when on_validate was used, which was fixed by updating them outside an if statement.