Adjusted VaR display section (added button) and got it to change to 999
I have changed the VaR section being displayed at the bottom right of the application so that it has a generate button next to it. I also have it set up so that when you press the VaR button, it will run a function that I've defined within the application, in this instance just setting the Var text to be 999. Next I will have to implement code from the other command line based code and get it to work with the GUI.