Fixed resizing issues and changes search bar
I realised the issue I was having was with the application resizing everything depending on the monitor I was using to display it in. For consistency sake, I've used SetProcessDPIAware to keep it the same size (without ruining my layouts) on any hardware, and have adjusted my layouts to match this. I also realised that I'm not doing anything with my search bar, so I replaced it with a label saying what the list beneath it is.