Displayed Backtesting and Fixed my Tickers
I finalised my displaying of the back testing, by having it displayed to the right of the VaR, turns out my model back testing was not wrong, it was just had a large outlier for the one stock I was testing. I also found out why a lot of my tickers for the FTSE100 weren't working, after reading online, many are often denoted with a .L after the ticker, so I started adding those to all the ones that would initially fail, and all of them begun to work. So now all the stocks are selectable and everything else is working, so the application is looking complete, minus the use of some input verification