Finished Initial Gui, implemented text and stock Verification
I think the GUI is now in a finished state for this term. It has most of the functionality that I originally intended, and looks good enough. I also added verification to the text boxes on the right, whenever you put a bad value in, it will set the current values to default, but if its a valid value that's too large, it will be set a max value that I have predetermined. I have also fixed 2 errors that would happen with the stocks code, comments in the file will explain what was happening. Finally, I've added appropriate comments, and neatened up the code so I am satisfied.