Generated Exe and fixed IDT name
I used pyinstaller to create an exe file for the application, I've tried to do it without it launching console but it seems to break every time I try and do it that way. So I've left it so that console comes up as well. I also realised I hadn't name the file Initial Test Design with a "'.py" so I fixed that as well.