Almost finished Documentation, renamed folders and moved files
I have almost completed my documentation, with me just needing to update my final Diary entry that will be appended to the end of the report. I also have neatened up my Git folders and files a bit more, so its easier to understand and navigate in the future. Everything now will be done to help finalise what I need done for the end of this term.