Final Push before Interim Deadline
I created a folder for my submission, where I added all the necessary files, then compiled it into a zip. I also added the final entry for this term for my diary. I also made 1 minor addition to the readMe and documentation. I cannot upload/push the .zip file as it appears to be too large, but it is just the same as what is in the folder of the same name, found in the same directory.